+++ grails-app/domain/com/cc/content/page/PageRevision.groovy
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2011, CauseCode Technologies Pvt Ltd, India.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or
+ * without modification, are not permitted.
+ */
+
+package com.cc.content.page
+
+import com.cc.content.ContentRevision
+
+class PageRevision extends ContentRevision {
+
+}