+++ test/unit/com/cc/content/ContentRevisionControllerTests.groovy
@@ -1,0 +1,17 @@
+package com.cc.content
+
+
+
+import grails.test.mixin.*
+import org.junit.*
+
+/**
+ * See the API for {@link grails.test.mixin.web.ControllerUnitTestMixin} for usage instructions
+ */
+@TestFor(ContentRevisionController)
+class ContentRevisionControllerTests {
+
+    void testSomething() {
+       fail "Implement me"
+    }
+}